giveFeedback()
This function giveFeedback() checks a condition tested in incrementCounter() that returns a true or false value to the variabled 'finished' in the constructor. 
If finished in the state class is 'true', it ends the quiz and gives a feedback to the user. This feedback contains the points gotten by the user, percentage of questions missed and the questions missed with their respective solutions