displayAnswer()
this.state.count is declared a counter to keep track of the start button. The start button is only active once so this condition calls count from the state and if the value is 0, it makes the start button active and it is more than 0, the start is disabled till the quiz ends.