displayAnswer() Line 227-244
This condition checks for if the answer given is correct and adds points if it is or if it is wrong and not add any point. If the answer picked by the user is wrong, the id of the question is loaded into the missedQuestions variable declared inside the state constructor.