Changing timer
the time countdown function has been changed. New code is on line 119 - 144. The counter is activated on click of the next/start button and onclick of any option, the timer stops. On countDown finish of the time, the options are disabled and the next button is activated. Also a time run out word is set into the state which is called below the timer display on the app.